+++ plugin-dev/Source/Sentry/Private/Apple/AppleSentrySubsystem.cpp
@@ -2,15 +2,15 @@
 
 #include "AppleSentrySubsystem.h"
 
-#include "SentryBreadcrumbApple.h"
-#include "SentryEventApple.h"
-#include "SentryScopeApple.h"
-#include "SentryUserApple.h"
-#include "SentryUserFeedbackApple.h"
-#include "SentryTransactionApple.h"
-#include "SentryTransactionContextApple.h"
-#include "SentrySamplingContextApple.h"
-#include "SentryIdApple.h"
+#include "AppleSentryBreadcrumb.h"
+#include "AppleSentryEvent.h"
+#include "AppleSentryScope.h"
+#include "AppleSentryUser.h"
+#include "AppleSentryUserFeedback.h"
+#include "AppleSentryTransaction.h"
+#include "AppleSentryTransactionContext.h"
+#include "AppleSentrySamplingContext.h"
+#include "AppleSentryId.h"
 
 #include "SentryBreadcrumb.h"
 #include "SentryEvent.h"
@@ -21,10 +21,13 @@
 #include "SentrySamplingContext.h"
 #include "SentryTraceSampler.h"
 
-#include "Infrastructure/SentryConvertersApple.h"
+#include "Infrastructure/AppleSentryConverters.h"
 
-#include "Convenience/SentryInclude.h"
-#include "Convenience/SentryMacro.h"
+#include "Convenience/AppleSentryInclude.h"
+#include "Convenience/AppleSentryMacro.h"
+
+#include "Utils/SentryFileUtils.h"
+#include "Utils/SentryLogUtils.h"
 
 #include "GenericPlatform/GenericPlatformOutputDevices.h"
 #include "HAL/FileManager.h"
@@ -34,10 +37,12 @@
 #include "Misc/Paths.h"
 #include "UObject/GarbageCollection.h"
 #include "UObject/UObjectThreadContext.h"
-#include "Utils/SentryLogUtils.h"
 
 void FAppleSentrySubsystem::InitWithSettings(const USentrySettings* settings, USentryBeforeSendHandler* beforeSendHandler, USentryBeforeBreadcrumbHandler* beforeBreadcrumbHandler, USentryTraceSampler* traceSampler)
 {
+	isScreenshotAttachmentEnabled = settings->AttachScreenshot;
+	isGameLogAttachmentEnabled = settings->EnableAutoLogAttachment;
+
 	[SENTRY_APPLE_CLASS(PrivateSentrySDKOnly) setSdkName:@"sentry.cocoa.unreal"];
 
 	dispatch_group_t sentryDispatchGroup = dispatch_group_create();
@@ -65,24 +70,18 @@ void FAppleSentrySubsystem::InitWithSettings(const USentrySettings* settings, US
 #if SENTRY_UIKIT_AVAILABLE
 			options.attachScreenshot = settings->AttachScreenshot;
 #endif
-			options.initialScope = ^(SentryScope* scope) {
-				if(settings->EnableAutoLogAttachment) {
-					const FString logFilePath = IFileManager::Get().ConvertToAbsolutePathForExternalAppForRead(*FGenericPlatformOutputDevices::GetAbsoluteLogFilename());
-					SentryAttachment* logAttachment = [[SENTRY_APPLE_CLASS(SentryAttachment) alloc] initWithPath:logFilePath.GetNSString()];
-					[scope addAttachment:logAttachment];
-				}
-				return scope;
-			};
 			options.onCrashedLastRun = ^(SentryEvent* event) {
 				if (settings->AttachScreenshot)
 				{
 					// If a screenshot was captured during assertion/crash in the previous app run
 					// find the most recent one and upload it to Sentry.
-					const FString& screenshotPath = GetLatestScreenshot();
-					if (!screenshotPath.IsEmpty())
-					{
-						UploadScreenshotForEvent(MakeShareable(new SentryIdApple(event.eventId)), screenshotPath);
-					}
+					UploadScreenshotForEvent(MakeShareable(new FAppleSentryId(event.eventId)), GetLatestScreenshot());
+				}
+				if(settings->EnableAutoLogAttachment)
+				{
+					// Unreal creates game log backups automatically on every app run. If logging is enabled for current configuration, SDK can
+					// find the most recent one and upload it to Sentry.
+					UploadGameLogForEvent(MakeShareable(new FAppleSentryId(event.eventId)), GetLatestGameLog());
 				}
 			};
 			options.beforeSend = ^SentryEvent* (SentryEvent* event) {
@@ -101,7 +100,7 @@ void FAppleSentrySubsystem::InitWithSettings(const USentrySettings* settings, US
 					return event;
 				}
 
-				USentryEvent* EventToProcess = USentryEvent::Create(MakeShareable(new SentryEventApple(event)));
+				USentryEvent* EventToProcess = USentryEvent::Create(MakeShareable(new FAppleSentryEvent(event)));
 
 				USentryEvent* ProcessedEvent = beforeSendHandler->HandleBeforeSend(EventToProcess, nullptr);
 
@@ -124,7 +123,7 @@ void FAppleSentrySubsystem::InitWithSettings(const USentrySettings* settings, US
 			{
 				options.tracesSampler = ^NSNumber* (SentrySamplingContext* samplingContext) {
 					FGCScopeGuard GCScopeGuard;
-					USentrySamplingContext* Context = USentrySamplingContext::Create(MakeShareable(new SentrySamplingContextApple(samplingContext)));
+					USentrySamplingContext* Context = USentrySamplingContext::Create(MakeShareable(new FAppleSentrySamplingContext(samplingContext)));
 					float samplingValue;
 					return traceSampler->Sample(Context, samplingValue) ? [NSNumber numberWithFloat:samplingValue] : nil;
 				};
@@ -146,7 +145,7 @@ void FAppleSentrySubsystem::InitWithSettings(const USentrySettings* settings, US
 						return breadcrumb;
 					}
 
-					USentryBreadcrumb* BreadcrumbToProcess = USentryBreadcrumb::Create(MakeShareable(new SentryBreadcrumbApple(breadcrumb)));
+					USentryBreadcrumb* BreadcrumbToProcess = USentryBreadcrumb::Create(MakeShareable(new FAppleSentryBreadcrumb(breadcrumb)));
 
 					USentryBreadcrumb* ProcessedBreadcrumb = beforeBreadcrumbHandler->HandleBeforeBreadcrumb(BreadcrumbToProcess, nullptr);
 
@@ -180,14 +179,14 @@ ESentryCrashedLastRun FAppleSentrySubsystem::IsCrashedLastRun()
 
 void FAppleSentrySubsystem::AddBreadcrumb(TSharedPtr<ISentryBreadcrumb> breadcrumb)
 {
-	TSharedPtr<SentryBreadcrumbApple> breadcrumbIOS = StaticCastSharedPtr<SentryBreadcrumbApple>(breadcrumb);
+	TSharedPtr<FAppleSentryBreadcrumb> breadcrumbIOS = StaticCastSharedPtr<FAppleSentryBreadcrumb>(breadcrumb);
 
 	[SENTRY_APPLE_CLASS(SentrySDK) addBreadcrumb:breadcrumbIOS->GetNativeObject()];
 }
 
 void FAppleSentrySubsystem::AddBreadcrumbWithParams(const FString& Message, const FString& Category, const FString& Type, const TMap<FString, FString>& Data, ESentryLevel Level)
 {
-	TSharedPtr<SentryBreadcrumbApple> breadcrumbIOS = MakeShareable(new SentryBreadcrumbApple());
+	TSharedPtr<FAppleSentryBreadcrumb> breadcrumbIOS = MakeShareable(new FAppleSentryBreadcrumb());
 	breadcrumbIOS->SetMessage(Message);
 	breadcrumbIOS->SetCategory(Category);
 	breadcrumbIOS->SetType(Type);
@@ -206,40 +205,49 @@ void FAppleSentrySubsystem::ClearBreadcrumbs()
 
 TSharedPtr<ISentryId> FAppleSentrySubsystem::CaptureMessage(const FString& message, ESentryLevel level)
 {
-	SentryId* id = [SENTRY_APPLE_CLASS(SentrySDK) captureMessage:message.GetNSString() withScopeBlock:^(SentryScope* scope){
-		[scope setLevel:SentryConvertersApple::SentryLevelToNative(level)];
-	}];
-
-	return MakeShareable(new SentryIdApple(id));
+	FSentryScopeDelegate onConfigureScope;
+	return CaptureMessageWithScope(message, onConfigureScope, level);
 }
 
 TSharedPtr<ISentryId> FAppleSentrySubsystem::CaptureMessageWithScope(const FString& message, const FSentryScopeDelegate& onConfigureScope, ESentryLevel level)
 {
-	SentryId* id = [SENTRY_APPLE_CLASS(SentrySDK) captureMessage:message.GetNSString() withScopeBlock:^(SentryScope* scope){
-		[scope setLevel:SentryConvertersApple::SentryLevelToNative(level)];
-		onConfigureScope.ExecuteIfBound(MakeShareable(new SentryScopeApple(scope)));
+	SentryId* nativeId = [SENTRY_APPLE_CLASS(SentrySDK) captureMessage:message.GetNSString() withScopeBlock:^(SentryScope* scope){
+		[scope setLevel:FAppleSentryConverters::SentryLevelToNative(level)];
+		onConfigureScope.ExecuteIfBound(MakeShareable(new FAppleSentryScope(scope)));
 	}];
 
-	return MakeShareable(new SentryIdApple(id));
+	TSharedPtr<ISentryId> id = MakeShareable(new FAppleSentryId(nativeId));
+
+	if (isGameLogAttachmentEnabled)
+	{
+		UploadGameLogForEvent(id, GetGameLogPath());
+	}
+
+	return id;
 }
 
 TSharedPtr<ISentryId> FAppleSentrySubsystem::CaptureEvent(TSharedPtr<ISentryEvent> event)
 {
-	TSharedPtr<SentryEventApple> eventIOS = StaticCastSharedPtr<SentryEventApple>(event);
-
-	SentryId* id = [SENTRY_APPLE_CLASS(SentrySDK) captureEvent:eventIOS->GetNativeObject()];
-	return MakeShareable(new SentryIdApple(id));
+	FSentryScopeDelegate onConfigureScope;
+	return CaptureEventWithScope(event,	onConfigureScope);
 }
 
 TSharedPtr<ISentryId> FAppleSentrySubsystem::CaptureEventWithScope(TSharedPtr<ISentryEvent> event, const FSentryScopeDelegate& onConfigureScope)
 {
-	TSharedPtr<SentryEventApple> eventIOS = StaticCastSharedPtr<SentryEventApple>(event);
+	TSharedPtr<FAppleSentryEvent> eventApple = StaticCastSharedPtr<FAppleSentryEvent>(event);
 
-	SentryId* id = [SENTRY_APPLE_CLASS(SentrySDK) captureEvent:eventIOS->GetNativeObject() withScopeBlock:^(SentryScope* scope) {
-		onConfigureScope.ExecuteIfBound(MakeShareable(new SentryScopeApple(scope)));
+	SentryId* nativeId = [SENTRY_APPLE_CLASS(SentrySDK) captureEvent:eventApple->GetNativeObject() withScopeBlock:^(SentryScope* scope) {
+		onConfigureScope.ExecuteIfBound(MakeShareable(new FAppleSentryScope(scope)));
 	}];
 
-	return MakeShareable(new SentryIdApple(id));
+	TSharedPtr<ISentryId> id = MakeShareable(new FAppleSentryId(nativeId));
+
+	if (isGameLogAttachmentEnabled)
+	{
+		UploadGameLogForEvent(id, GetGameLogPath());
+	}
+
+	return id;
 }
 
 TSharedPtr<ISentryId> FAppleSentrySubsystem::CaptureEnsure(const FString& type, const FString& message)
@@ -251,21 +259,29 @@ TSharedPtr<ISentryId> FAppleSentrySubsystem::CaptureEnsure(const FString& type, 
 	SentryEvent *exceptionEvent = [[SENTRY_APPLE_CLASS(SentryEvent) alloc] init];
 	exceptionEvent.exceptions = nativeExceptionArray;
 
-	SentryId* id = [SENTRY_APPLE_CLASS(SentrySDK) captureEvent:exceptionEvent];
-	return MakeShareable(new SentryIdApple(id));
+	SentryId* nativeId = [SENTRY_APPLE_CLASS(SentrySDK) captureEvent:exceptionEvent];
+
+	TSharedPtr<ISentryId> id = MakeShareable(new FAppleSentryId(nativeId));
+
+	if (isGameLogAttachmentEnabled)
+	{
+		UploadGameLogForEvent(id, GetGameLogPath());
+	}
+
+	return id;
 }
 
 
 void FAppleSentrySubsystem::CaptureUserFeedback(TSharedPtr<ISentryUserFeedback> userFeedback)
 {
-	TSharedPtr<SentryUserFeedbackApple> userFeedbackIOS = StaticCastSharedPtr<SentryUserFeedbackApple>(userFeedback);
+	TSharedPtr<FAppleSentryUserFeedback> userFeedbackIOS = StaticCastSharedPtr<FAppleSentryUserFeedback>(userFeedback);
 
 	[SENTRY_APPLE_CLASS(SentrySDK) captureFeedback:userFeedbackIOS->GetNativeObject()];
 }
 
 void FAppleSentrySubsystem::SetUser(TSharedPtr<ISentryUser> user)
 {
-	TSharedPtr<SentryUserApple> userIOS = StaticCastSharedPtr<SentryUserApple>(user);
+	TSharedPtr<FAppleSentryUser> userIOS = StaticCastSharedPtr<FAppleSentryUser>(user);
 
 	[SENTRY_APPLE_CLASS(SentrySDK) setUser:userIOS->GetNativeObject()];
 }
@@ -278,14 +294,14 @@ void FAppleSentrySubsystem::RemoveUser()
 void FAppleSentrySubsystem::ConfigureScope(const FSentryScopeDelegate& onConfigureScope)
 {
 	[SENTRY_APPLE_CLASS(SentrySDK) configureScope:^(SentryScope* scope) {
-		onConfigureScope.ExecuteIfBound(MakeShareable(new SentryScopeApple(scope)));
+		onConfigureScope.ExecuteIfBound(MakeShareable(new FAppleSentryScope(scope)));
 	}];
 }
 
 void FAppleSentrySubsystem::SetContext(const FString& key, const TMap<FString, FString>& values)
 {
 	[SENTRY_APPLE_CLASS(SentrySDK) configureScope:^(SentryScope* scope) {
-		[scope setContextValue:SentryConvertersApple::StringMapToNative(values) forKey:key.GetNSString()];
+		[scope setContextValue:FAppleSentryConverters::StringMapToNative(values) forKey:key.GetNSString()];
 	}];
 }
 
@@ -306,7 +322,7 @@ void FAppleSentrySubsystem::RemoveTag(const FString& key)
 void FAppleSentrySubsystem::SetLevel(ESentryLevel level)
 {
 	[SENTRY_APPLE_CLASS(SentrySDK) configureScope:^(SentryScope* scope) {
-		[scope setLevel:SentryConvertersApple::SentryLevelToNative(level)];
+		[scope setLevel:FAppleSentryConverters::SentryLevelToNative(level)];
 	}];
 }
 
@@ -324,16 +340,16 @@ TSharedPtr<ISentryTransaction> FAppleSentrySubsystem::StartTransaction(const FSt
 {
 	id<SentrySpan> transaction = [SENTRY_APPLE_CLASS(SentrySDK) startTransactionWithName:name.GetNSString() operation:operation.GetNSString()];
 
-	return MakeShareable(new SentryTransactionApple(transaction));
+	return MakeShareable(new FAppleSentryTransaction(transaction));
 }
 
 TSharedPtr<ISentryTransaction> FAppleSentrySubsystem::StartTransactionWithContext(TSharedPtr<ISentryTransactionContext> context)
 {
-	TSharedPtr<SentryTransactionContextApple> transactionContextIOS = StaticCastSharedPtr<SentryTransactionContextApple>(context);
+	TSharedPtr<FAppleSentryTransactionContext> transactionContextIOS = StaticCastSharedPtr<FAppleSentryTransactionContext>(context);
 
 	id<SentrySpan> transaction = [SENTRY_APPLE_CLASS(SentrySDK) startTransactionWithContext:transactionContextIOS->GetNativeObject()];
 
-	return MakeShareable(new SentryTransactionApple(transaction));
+	return MakeShareable(new FAppleSentryTransaction(transaction));
 }
 
 TSharedPtr<ISentryTransaction> FAppleSentrySubsystem::StartTransactionWithContextAndTimestamp(TSharedPtr<ISentryTransactionContext> context, int64 timestamp)
@@ -344,12 +360,12 @@ TSharedPtr<ISentryTransaction> FAppleSentrySubsystem::StartTransactionWithContex
 
 TSharedPtr<ISentryTransaction> FAppleSentrySubsystem::StartTransactionWithContextAndOptions(TSharedPtr<ISentryTransactionContext> context, const TMap<FString, FString>& options)
 {
-	TSharedPtr<SentryTransactionContextApple> transactionContextIOS = StaticCastSharedPtr<SentryTransactionContextApple>(context);
+	TSharedPtr<FAppleSentryTransactionContext> transactionContextIOS = StaticCastSharedPtr<FAppleSentryTransactionContext>(context);
 
 	id<SentrySpan> transaction = [SENTRY_APPLE_CLASS(SentrySDK) startTransactionWithContext:transactionContextIOS->GetNativeObject()
-		customSamplingContext:SentryConvertersApple::StringMapToNative(options)];
+		customSamplingContext:FAppleSentryConverters::StringMapToNative(options)];
 
-	return MakeShareable(new SentryTransactionApple(transaction));
+	return MakeShareable(new FAppleSentryTransaction(transaction));
 }
 
 TSharedPtr<ISentryTransactionContext> FAppleSentrySubsystem::ContinueTrace(const FString& sentryTrace, const TArray<FString>& baggageHeaders)
@@ -386,38 +402,61 @@ TSharedPtr<ISentryTransactionContext> FAppleSentrySubsystem::ContinueTrace(const
 
 	// currently `sentry-cocoa` doesn't have API for `SentryTransactionContext` to set `baggageHeaders`
 
-	return MakeShareable(new SentryTransactionContextApple(transactionContext));
+	return MakeShareable(new FAppleSentryTransactionContext(transactionContext));
 }
 
-void FAppleSentrySubsystem::UploadScreenshotForEvent(TSharedPtr<ISentryId> eventId, const FString& screenshotPath) const
+void FAppleSentrySubsystem::UploadAttachmentForEvent(TSharedPtr<ISentryId> eventId, const FString& filePath, const FString& name, bool deleteAfterUpload) const
 {
 	IFileManager& fileManager = IFileManager::Get();
-	if (!fileManager.FileExists(*screenshotPath))
+	if (!fileManager.FileExists(*filePath))
 	{
-		UE_LOG(LogSentrySdk, Error, TEXT("Failed to upload screenshot - path provided did not exist: %s"), *screenshotPath);
+		UE_LOG(LogSentrySdk, Error, TEXT("Failed to upload attachment - file path provided did not exist: %s"), *filePath);
 		return;
 	}
 
-	const FString& screenshotFilePathExt = fileManager.ConvertToAbsolutePathForExternalAppForRead(*screenshotPath);
+	const FString& filePathExt = fileManager.ConvertToAbsolutePathForExternalAppForRead(*filePath);
 
-	SentryAttachment* screenshotAttachment = [[SENTRY_APPLE_CLASS(SentryAttachment) alloc] initWithPath:screenshotFilePathExt.GetNSString() filename:@"screenshot.png"];
+	SentryAttachment* attachment = [[SENTRY_APPLE_CLASS(SentryAttachment) alloc] initWithPath:filePathExt.GetNSString() filename:name.GetNSString()];
 
 	SentryOptions* options = [SENTRY_APPLE_CLASS(PrivateSentrySDKOnly) options];
 	int32 size = options.maxAttachmentSize;
 
-	SentryEnvelopeItem* envelopeItem = [[SENTRY_APPLE_CLASS(SentryEnvelopeItem) alloc] initWithAttachment:screenshotAttachment maxAttachmentSize:size];
+	SentryEnvelopeItem* envelopeItem = [[SENTRY_APPLE_CLASS(SentryEnvelopeItem) alloc] initWithAttachment:attachment maxAttachmentSize:size];
 
-	SentryId* id = StaticCastSharedPtr<SentryIdApple>(eventId)->GetNativeObject();
+	SentryId* id = StaticCastSharedPtr<FAppleSentryId>(eventId)->GetNativeObject();
 
 	SentryEnvelope* envelope = [[SENTRY_APPLE_CLASS(SentryEnvelope) alloc] initWithId:id singleItem:envelopeItem];
 
 	[SENTRY_APPLE_CLASS(PrivateSentrySDKOnly) captureEnvelope:envelope];
 
-	// After uploading screenshot it's no longer needed so delete
-	if (!fileManager.Delete(*screenshotPath))
+	if (deleteAfterUpload)
 	{
-		UE_LOG(LogSentrySdk, Error, TEXT("Failed to delete screenshot: %s"), *screenshotPath);
+		if (!fileManager.Delete(*filePath))
+		{
+			UE_LOG(LogSentrySdk, Error, TEXT("Failed to delete file attachment after upload: %s"), *filePath);
+		}
 	}
+}
+
+void FAppleSentrySubsystem::UploadScreenshotForEvent(TSharedPtr<ISentryId> eventId, const FString& screenshotPath) const
+{
+	if (screenshotPath.IsEmpty())
+	{
+		// Screenshot capturing is a best-effort solution so if one wasn't captured (path is empty) skip the upload
+		return;
+	}
+
+	UploadAttachmentForEvent(eventId, screenshotPath, TEXT("screenshot.png"), true);
+}
+
+void FAppleSentrySubsystem::UploadGameLogForEvent(TSharedPtr<ISentryId> eventId, const FString& logFilePath) const
+{
+#if NO_LOGGING
+	// If writing logs to a file is disabled (i.e. default behavior for Shipping builds) skip the upload
+	return;
+#endif
+
+	UploadAttachmentForEvent(eventId, logFilePath, SentryFileUtils::GetGameLogName());
 }
 
 FString FAppleSentrySubsystem::GetScreenshotPath() const