+++ plugin-dev/Source/Sentry/Private/HAL/PlatformSentryUser.h
@@ -1,24 +1,16 @@
 #pragma once
 
 #if PLATFORM_ANDROID
-#include "Android/SentryUserAndroid.h"
+#include "Android/AndroidSentryUser.h"
 #elif PLATFORM_APPLE
-#include "Apple/SentryUserApple.h"
+#include "Apple/AppleSentryUser.h"
 #elif USE_SENTRY_NATIVE
 #include "GenericPlatform/GenericPlatformSentryUser.h"
 #else
-#include "Interface/SentryUserInterface.h"
+#include "Null/NullSentryUser.h"
 #endif
 
 static TSharedPtr<ISentryUser> CreateSharedSentryUser()
 {
-#if PLATFORM_ANDROID
-	return MakeShareable(new SentryUserAndroid);
-#elif PLATFORM_APPLE
-	return MakeShareable(new SentryUserApple);
-#elif USE_SENTRY_NATIVE
-	return MakeShareable(new FGenericPlatformSentryUser);
-#else
-	return nullptr;
-#endif
+	return MakeShareable(new FPlatformSentryUser);
 }